readme: Add ld.so.preload injection option
When the user program is spawned by another process automatically,
e.g., when using a Jupyter Notebook Server, it is not easy to
set the `LD_PRELOAD=libnvshare.so` env variable.

Libnvshare only activates when the program calls `cuInit()`. Thus,
it is safe to preload it to all programs, and the overhead is
negligible, especially when we have no other intuitive choice.

The dynamic linker/loader reads libraries to preload from
/etc/ld.so.preload [1] in addition to the LD_PRELOAD env.

Add a third option in the "Local Usage" section with instructions on
appending libnvshare.so to `/etc/ld.so.preload`.

Also fix a small typo in the usage_k8s_conf section.

[1]: https://man7.org/linux/man-pages/man8/ld.so.8.html

Signed-off-by: George Alexopoulos <giorgosalexo0@gmail.com>